+++ apu/hermite_resampler.h
@@ -80,8 +80,8 @@ class HermiteResampler : public Resampler
 
                 while (r_frac <= 1.0 && o_position < num_samples)
                 {
-                    hermite_val[0] = hermite (r_frac, r_left [0], r_left [1], r_left [2], r_left [3]);
-                    hermite_val[1] = hermite (r_frac, r_right[0], r_right[1], r_right[2], r_right[3]); 
+                    hermite_val[0] = hermite (r_frac, (float)r_left [0], (float)r_left [1], (float)r_left [2], (float)r_left [3]);
+                    hermite_val[1] = hermite (r_frac, (float)r_right[0], (float)r_right[1], (float)r_right[2], (float)r_right[3]); 
                     data[o_position]     = SHORT_CLAMP (hermite_val[0]);
                     data[o_position + 1] = SHORT_CLAMP (hermite_val[1]);
